refactor: split proposals into multiple files
