+++ governance/xc_admin/packages/xc_admin_frontend/components/tabs/Proposals/utils.ts
@@ -1,0 +1,17 @@
+import { MultisigAccount, TransactionAccount } from '@sqds/mesh/lib/types'
+
+
+export const getProposalStatus = (
+    proposal: TransactionAccount | undefined,
+    multisig: MultisigAccount | undefined
+): string => {
+    if (multisig && proposal) {
+        const onChainStatus = Object.keys(proposal.status)[0]
+        return proposal.transactionIndex <= multisig.msChangeIndex &&
+            (onChainStatus == 'active' || onChainStatus == 'draft')
+            ? 'expired'
+            : onChainStatus
+    } else {
+        return 'unkwown'
+    }
+}
